requirements.txt
aioredis==1.3.1
amqp==5.0.5
asgiref==3.3.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
billiard==3.6.3.0
celery==5.0.5
certifi==2018.1.18
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
dj-database-url==0.5.0
Django==3.1.7
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-heroku==0.3.1
django-redis==4.12.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-timezone-field==4.1.1
djangochannelsrestframework==0.2.1
djangorestframework==3.12.2
hiredis==1.1.0
httplib2==0.9.2
hyperlink==21.0.0
idna==2.6
importlib-metadata==3.7.3
incremental==21.3.0
keyring==10.6.0
keyrings.alt==3.0
kombu==5.0.2
msgpack==1.0.2
netifaces==0.10.4
olefile==0.45.1
pexpect==4.2.1
prompt-toolkit==3.0.17
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pyinotify==0.9.6
pyOpenSSL==20.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2021.1
pyxdg==0.25
redis==3.5.3
reportlab==3.4.0
requests==2.18.4
SecretStorage==2.3.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
urllib3==1.22
wcwidth==0.2.5
websocket-client==0.44.0
whitenoise==5.2.0
twilio==6.62.1
django-environ==0.4.5
mysqlclient==2.0.3
django-storages==1.11.1
# botocore==1.22.4
django-taggit==1.5.1
django-taggit-serializer==0.1.7
